backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
sqlalchemy==2.0.27
pydantic==2.6.1
pydantic-settings==2.1.0
alembic==1.13.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9

celery==5.3.6
redis==5.0.1
google-generativeai==0.4.0
python-dotenv==1.0.1
pytest==8.0.0
httpx==0.26.0

pypdf==5.4.0
chromadb==0.4.22
langchain==0.1.9
langchain-google-genai==0.0.9
